+++ src/Symfony/Component/Uid/CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 ---
 
  * Add `UuidV1::toV6()`, `UuidV1::toV7()` and `UuidV6::toV7()`
+ * Add `AbstractUid::toString()`
 
 6.2
 ---